+++ package.json
@@ -13,10 +13,5 @@
     "@types/react-dom": "16.9.5",
     "typescript": "3.7.5"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "browserslist": [">0.2%", "not dead", "not ie <= 11", "not op_mini all"]
 }
