--- src/App.mod.css
@@ -1,15 +1,0 @@
-.container {
-	max-width: 600px;
-	margin: auto;
-}
-
-.dialog {
-	border: 1px solid #777;
-	border-radius: 4px;
-	padding: 8px;
-	margin: 8px 0;
-}
-
-p {
-	line-height: 1.4;
-}
